requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
docutils==0.17.1
psycopg2==2.9.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
whitenoise==5.3.0
